# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is API Documentation for Product Service Services.",
        "title": "Product Service API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/health-check": {
            "get": {
                "description": "LOV untuk health check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HealthCheck"
                ],
                "summary": "Get Health Check",
                "responses": {}
            }
        },
        "/user/login": {
            "post": {
                "description": "Login For User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "API Login User",
                "parameters": [
                    {
                        "description": "Request Parameters",
                        "name": "LoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userModel.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Success",
                        "schema": {
                            "$ref": "#/definitions/userModel.CreateUserResponse"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Create For User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "API Create User",
                "parameters": [
                    {
                        "description": "Request Parameters",
                        "name": "CreateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userModel.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Success",
                        "schema": {
                            "$ref": "#/definitions/userModel.CreateUserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "userModel.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone_number"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "userModel.CreateUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "userModel.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}